// lib/api/generated/model/wrapperResponseLoginCommandResponseDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Immoplus
 * Immoplus
 * OpenAPI spec version: 1.0
 */
import type { LoginCommandResponse } from './loginCommandResponse';

export interface WrapperResponseLoginCommandResponseDto {
  data: LoginCommandResponse;
  currentPage: number;
  totalPages: number;
  pageSize: number;
  totalCount: number;
  hasPrevious: boolean;
  hasNext: boolean;
}